requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
idna==3.4
Pillow==9.2.0
requests==2.28.1
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.12
